pyproject.toml
[tool.poetry]
name = "store-visit-backend"
version = "0.1.0"
description = "An API for a mobile application where a customer's field employee will conduct visits to stores."
authors = ["amantur02 <amantur.ubaid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
pydantic = "^2.3.0"
sqlalchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
uvicorn = "^0.23.2"
gunicorn = "^21.2.0"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.7"
passlib = "^1.7.4"
asyncpg = "^0.28.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
